chore: rename test files
